+++ src/tightdb/column_string.cpp
@@ -216,6 +216,13 @@ StringIndex* AdaptiveStringColumn::create_search_index()
 }
 
 
+void AdaptiveStringColumn::destroy_search_index() TIGHTDB_NOEXCEPT
+{
+    delete m_search_index;
+    m_search_index = 0;
+}
+
+
 void AdaptiveStringColumn::set_search_index_ref(ref_type ref, ArrayParent* parent,
                                                 size_t ndx_in_parent, bool allow_duplicate_valaues)
 {