+++ src/tightdb/column_string.hpp
@@ -98,9 +98,12 @@ public:
     const StringIndex* get_search_index() const TIGHTDB_NOEXCEPT;
     StringIndex* release_search_index() TIGHTDB_NOEXCEPT;
     StringIndex* create_search_index();
+
     // Simply inserts all column values in the index in a loop
     void populate_search_index();
+    void destroy_search_index() TIGHTDB_NOEXCEPT TIGHTDB_OVERRIDE;
 
+    // Optimizing data layout
     // Optimizing data layout. enforce == true will enforce enumeration; 
     // enforce == false will auto-evaluate if it should be enumerated or not
     bool auto_enumerate(ref_type& keys, ref_type& values, bool enforce = false) const;