+++ src/tightdb/exceptions.hpp
@@ -99,6 +99,10 @@ public:
         // already had a primary key.
         has_primary_key,
 
+        /// Indicates that a modification to a column was attempted that cannot
+        /// be done because the column is the primary key of the table.
+        is_primary_key,
+
         /// Indicates that a modification was attempted that would have produced a
         /// duplicate primary value.
         unique_constraint_violation,