+++ src/tightdb/group_shared.hpp
@@ -600,10 +600,20 @@ inline SharedGroup::SharedGroup(const std::string& file, bool no_create, Durabil
     m_group(Group::shared_tag())
 {
     open(file, no_create, dlevel, false, key);
+
     // Upgrade file format from 2 to 3 (no-op if already 3)
-    begin_write();
-    m_group.upgrade_file_format();
-    commit();
+    bool upgrade = false;
+    begin_read();
+    upgrade = m_group.get_file_format() < default_file_format_version;
+    end_read();    
+
+    // Only create write transaction if needed; that's why we test whether to upgrade or not in a separate read 
+    // transaction. Else unit tests would fail.
+    if (upgrade) {
+        begin_write();
+        m_group.upgrade_file_format();
+        commit();
+    }
 }
 
 inline SharedGroup::SharedGroup(unattached_tag) TIGHTDB_NOEXCEPT: